+++ src/SAOImageDS9.jl
@@ -7,16 +7,30 @@
 #------------------------------------------------------------------------------
 #
 # This file is part of SAOImageDS9.jl released under the MIT "expat" license.
-# Copyright (C) 2016-2020, Éric Thiébaut (https://github.com/emmt).
+# Copyright (c) 2016-2025, Éric Thiébaut (https://github.com/emmt).
 #
 
 module SAOImageDS9
+
+export ds9accesspoint
+export ds9connect
+export ds9cursor
+export ds9disconnect
+export ds9draw
+export ds9get
+export ds9getregions
+export ds9launch
+export ds9message
+export ds9quit
+export ds9set
+export ds9wcs
 
 export DS9
 const DS9 = SAOImageDS9
 
 using XPA
-using XPA: AccessPoint, TupleOf, connection, join_arguments
+using XPA: AccessPoint, Shape, TupleOf, connection, join_arguments
+using XPA: preserve_state, restore_state
 
 using TwoDimensional
 
@@ -27,281 +41,248 @@ using Base: ENV
 using Base.Iterators: Pairs
 
 # FITS pixel types.
-const PixelTypes = Union{UInt8,Int16,Int32,Int64,Float32,Float64}
-const PIXELTYPES = (UInt8, Int16, Int32, Int64, Float32, Float64)
+const PIXEL_TYPES = (UInt8, Int16, Int32, Int64, Float32, Float64)
+const PixelTypes = Union{PIXEL_TYPES...,}
 
-"""
-    accesspoint::Uniont{XPA.AccessPoint,Nothing}
+#---------------------------------------------------------------------------- ACCESS-POINT -
 
-The current XPA access point to the SAOImage/DS9 server.
+# Private global variable storing the default access-point to SAOImage/DS9.
+const _DEFAULT_APT = Ref{AccessPoint}()
 
-This access point can be set by calling the [`SAOImageDS9.connect`](@ref)
-method. If no access point has been chosen this variable is set to `nothing`.
-To automatically connect to SAOImage/DS9 if not yet done, you can do:
-
-```julia
-if isnothing(SAOImageDS9.accesspoint)
-    SAOImageDS9.connect()
-end
-```
-
-See also [`SAOImageDS9.connect`](@ref).
-"""
-accesspoint::Union{AccessPoint,Nothing} = nothing
-
-"""
-    _apt()
-
-The default access point.
-
-This function just returns [`access_point`](@ref) if this variable is set, or
-establishes a new connection using [`connect`](@ref).
-"""
-function _apt()
-    global accesspoint
-    if isnothing(accesspoint)
-        try
-            accesspoint = connect()
-        catch
-            @warn """Failed to automatically connect to SAOImage/DS9.
-                Launch ds9 then do `ds9connect()`"""
-        end
-    end
-    accesspoint
-end
-
-
-"""
-    SAOImageDS9.connect(ident="DS9:*"; method="local") -> apt
-
-Set the access point for further SAOImage/DS9 commands.
-
-The argument `ident` identifies the XPA access point: it can be a template
-string like `"DS9:*"` which is the default value or a regular expression.  The
-returned value is the name of the access point.
-
-To retrieve the name of the current SAOImage/DS9 access point, call the
-[`SAOImageDS9.accesspoint`](@ref) method.
-
-"""
-function connect(ident::Union{Regex,AbstractString} = "DS9:*"; method="local", kwds...)
-    global accesspoint
-    if !haskey(ENV, "XPA_METHOD") && !isnothing(method)
-        ENV["XPA_METHOD"] = string(method)
-    end
-    apt = XPA.find(ident; kwds...)
-    apt === nothing && error("no matching SAOImage/DS9 server found")
-    rep = XPA.get(apt, "version"; nmax=1)
-    if length(rep) != 1 || ! XPA.verify(rep)
-        error("XPA server at address \"" * apt *
-              "\" does not seem to be a SAOImage/DS9 server")
-    end
-    accesspoint = apt
-end
-
-
-"""
-    ds9select(ident="DS9:*"; method="local"; silent=false, interactive=true)
-
-Select a DS9 window for further interactions.
-
-If multiple DS9 windows match the `ident`, the user can select the correct
-window using a simple interface; if `interactive=false` the first matching
-window is selected.
-
-!!! warning
-    If `method` is not `undefined`, the environment variable `XPA_METHOD` is set
-    to the corresponding string. Typically the most reliable connection is
-    `local`, the default value.
-"""
-function ds9select(ident::Union{Regex,AbstractString}="DS9:*"; method="local",
-    silent=false, interactive=true)
-    global accesspoint
-    if !haskey(ENV, "XPA_METHOD") && !isnothing(method)
-        ENV["XPA_METHOD"] = string(method)
-    end
-    i = findfirst(isequal(':'), ident)
-    if i === nothing
-        # allow any class
-        class = "*"
-        name = ident
-    else
-        class = ident[1:i-1]
-        name = ident[i+1:end]
-    end
-    apts = XPA.list()
-    good_apts = filter(a -> (name == "*" || name == a.name) && (class == "*" || class == a.class), apts)
-    if length(good_apts) == 0
-        if !silent
-            @error "No valid XPA access point found (server not reachable?)"
-        end
-        return nothing
-    else
-        if length(good_apts) == 1 || !interactive
-            choice = 1
-        else
-            menu = RadioMenu(["$(p.class):$(p.name) (user=$(p.user))" for p ∈ apts])
-            choice = request("Please select the correct access point:", menu)
-            if choice == -1
-                return nothing
-            end
-        end
-        if !silent
-            class = good_apts[choice].class
-            name = good_apts[choice].name
-            user = good_apts[choice].user
-            @info "Connected to the XPA access point $class:$name (user=$user)"
-        end
-        accesspoint = good_apts[choice]
-        return nothing
-    end
+# `default_apt()` yields the default access-point and attempts to automatically find one
+# if the default one is closed.
+function default_apt()
+    apt = ds9accesspoint()
+    return isopen(apt) ? apt : ds9connect()
 end
 
 """
-    ds9([name]; method="local", path="ds9")
+    ds9accesspoint() -> apt
 
-Launch the DS9 application.
+Yield the current default XPA access-point to SAOImage/DS9.
 
-By default the application will be named using the current PID.
+# See also
 
-The `method` optional keywords is used to set the XPA communication method:
-"local" is the recommended way for local executions.
+[`ds9connect`](@ref) and  [`ds9disconnect`](@ref).
 
-This function authomaticall sets the default access point, so that all further
-requests are forwarded to the newly open DS9 window.
 """
-function ds9(name::String=string(getpid()); method="local", path="ds9", silent=false)
-    global accesspoint
-    if !haskey(ENV, "XPA_METHOD") && !isnothing(method)
-        ENV["XPA_METHOD"] = string(method)
+ds9accesspoint() = _DEFAULT_APT[]
+
+"""
+    ds9disconnect()
+
+Forget the default XPA access-point to SAOImage/DS9.
+
+# See also
+
+[`ds9accesspoint`](@ref) and [`ds9connect`](@ref).
+
+"""
+ds9disconnect() = ds9connect(AccessPoint())
+
+"""
+    ds9connect(apt::XPA.AccessPoint) -> apt
+    ds9connect(addr::AbstractString) -> apt
+    ds9connect(; kwds...) -> apt
+    ds9connect(f; kwds...) -> apt
+
+Set the default XPA access-point to SAOImage/DS9 and return it. The access-point may be
+fully specified (1st example), specified by its address (2nd example), or found among a
+list of running SAOImage/DS9 (3rd and 4th examples) by calling `XPA.find` with filter
+function `f` and keywords `kwds...`. If not specified (3rd example), the default filter
+function is:
+
+   apt -> apt.class == "DS9" && apt.user == ENV["USER"]
+
+When calling `XPA.find`, the default value of the `select` keyword is `:interact` if Julia
+is running an interactive session.
+
+# See also
+
+[`ds9accesspoint`](@ref) and [`ds9disconnect`](@ref).
+
+"""
+function ds9connect(apt::XPA.AccessPoint)
+    _DEFAULT_APT[] = apt
+    return apt
+end
+
+function ds9connect(addr::AbstractString)
+    return ds9connect(AccessPoint(address=addr))
+end
+
+function ds9connect(; kwds...)
+    return ds9connect(; kwds...) do apt
+        global ENV
+        apt.class == "DS9" && apt.user == ENV["USER"]
     end
-    command = detach(`$path -xpa yes -xpa connect -title $name`)
-    run(command; wait=false)
-    if !silent
-        printstyled("[ Info: "; color=Base.default_color_info, bold=true)
-        print("Opening DS9")
+end
+
+function ds9connect(f; select = isinteractive() ? :interact : :throw, kwds...)
+    return ds9connect(XPA.find(f; select=select, kwds...))
+end
+
+#--------------------------------------------------------------------------- LAUNCH / QUIT -
+
+"""
+    ds9launch([name]; method="local", exe="ds9", timeout=30, quiet=false)
+
+Launch the SAOImage/DS9 application and connect to it. Optional argument `name` is the
+name (and title) of the SAOImage/DS9 server to identity it. By default, the name is the
+given by the current PID. The default access-point is set to that of the new server.
+
+# Keywords
+
+- `method` is the type of connection to use. The default method is `ENV["XPA_METHOD"]` if
+  this environment variable is set and `"local"` otherwise.
+
+- `exe` is the path to SAOImage/DS9 executable.
+
+- `timeout` is the maximum number of seconds to wait for connecting.
+
+- `quiet` specifies whether to print information and warning messages.
+
+# See also
+
+[`ds9connect`](@ref) and [`ds9quit`](@ref).
+
+"""
+function ds9launch(name::AbstractString = string(getpid());
+                   method::Union{Symbol,AbstractString} = get(ENV, "XPA_METHOD", "local"),
+                   exe="ds9",
+                   quiet::Bool=false, timeout::Real=30)
+    global ENV
+    preserve_state(ENV, "XPA_METHOD") do
+        delete!(ENV, "XPA_METHOD")
+        run(detach(`$exe -xpa $method -title $name`); wait=false)
     end
-    for i ∈ 1:20
-        silent || print(".")
+    if !quiet
+        printstyled("[ Info:"; color=Base.default_color_info, bold=true)
+        print(" Opening DS9 with name $(repr(name))")
+    end
+    maxtime = time() + timeout
+    while time() ≤ maxtime
+        quiet || print(".")
         sleep(0.4)
-        apt = XPA.find("DS9:$name")
+        apt = XPA.find(; method=method, select=:throw) do apt
+            apt.class == "DS9" && apt.name == name
+        end
         if !isnothing(apt)
-            accesspoint = apt
-            silent || println(" done")
-            return
+            ds9connect(apt)
+            quiet || println(" done")
+            return nothing
         end
     end
-    silent || println(" failed")
+    quiet || println(" failed")
     @warn "Timeout establishing an XPA connection."
+    return nothing
 end
 
+"""
+    ds9quit()
+    ds9quit(apt::XPA.Accesspoint)
+    ds9quit(addr::AbstractString)
+
+Require SAOImage/DS9 application to quit. With no argument, the default SAOImage/DS9
+application is closed. Otherwise, a specific SAOImage/DS9 application may be targeted by
+specifying its access-point `apt` or the address `addr` of its access-point.
+
+See also [`ds9accesspoint`](@ref) and  [`ds9disconnect`](@ref).
 
 """
-    get([accesspoint,] [T, [dims,]] args...)
+ds9quit(addr::AbstractString) = ds9quit(AccessPoint(address = addr))
 
-Send a "get" request to the SAOImage/DS9 server.
+function ds9quit(apt::AccessPoint = default_apt())
+    if isopen(apt)
+        ds9set(apt, "quit")
+        if apt.address == default_apt().address
+            ds9disconnect()
+        end
+    end
+    return nothing
+end
 
-The request is made of arguments `args...` converted into strings and merged
-with separating spaces.
+#------------------------------------------------------------------------------- GET / SET -
+
+"""
+    ds9get([apt,] [T, [dims,]] args...)
+
+Send a "get" request to a SAOImage/DS9 server.
+
+The request is made of arguments `args...` converted into strings and concatenated with
+separating spaces.
 
 The returned value depends on the optional arguments `T` and `dims`:
 
-* If neither `T` nor `dims` are specified, the output is converted using an
-  heuristic method to suitable scalar or vector.
+FIXME: * If neither `T` nor `dims` are specified, the output is converted using an heuristic method
+FIXME:   to suitable scalar or vector.
+
 * If only `T` is specified, it can be:
+
   - `String` to return the answer as a single string;
-  - `Vector{String}}` or `Tuple{Vararg{String}}` to return the answer split in
-    words as a vector or as a tuple of strings;
-  - `T` where `T<:Real` to return a value of type `T` obtained by parsing the
-    textual answer.
-  - `Tuple{Vararg{T}}` where `T<:Real` to return a value of type `T` obtained
-    by parsing the textual answer;
-  - `Vector{T}` where `T` is not `String` to return the binary contents of the
-    answer as a vector of type `T`;
-* If both `T` and `dims` are specified, `T` can be an array type like
-  `Array{S}` or `Array{S,N}` and `dims` a list of `N` dimensions to retrieve
-  the binary contents of the answer as an array of type `Array{S,N}`.
 
-See also: [`SAOImageDS9.connect`](@ref), [`SAOImageDS9.set`](@ref) and `XPA.get`.
+  - `Vector{String}}` or `Tuple{Vararg{String}}` to return the answer split in words as a
+    vector or as a tuple of strings;
+
+  - `T` where `T<:Real` to return a value of type `T` obtained by parsing the textual
+    answer.
+
+  - `Tuple{Vararg{T}}` where `T<:Real` to return a value of type `T` obtained by parsing the
+    textual answer;
+
+  - `Vector{T}` where `T` is not `String` to return the binary contents of the answer as a
+    vector of type `T`;
+
+* If both `T` and `dims` are specified, `T` can be an array type like `Array{S}` or
+  `Array{S,N}` and `dims` a list of `N` dimensions to retrieve the binary contents of the
+  answer as an array of type `Array{S,N}`.
+
+# See also
+
+[`ds9connect`](@ref) and [`ds9set`](@ref).
+
 """
-function get(ap::AccessPoint, args...; kw...)
-    command = join_arguments(args)
-    if string(command) != ""
-        r = XPA.get(ap, string(command); kw...)
-        if XPA.has_errors(r)
-            m = XPA.get_message(r)
-            msg = strip(m[10:end])
-            f = findfirst("(DS9:", msg)
-            if !isnothing(f)
-                msg = msg[begin:first(f)-1]
-            end
-            @warn "XPA $msg"
-        elseif r.replies == 0
-            @warn "No replies for command `$command`"
-        end
-        data = XPA.get_data(String, r)
-        lines = filter(!isempty, split(data, "\n"))
-        if isnothing(findfirst(x -> isnothing(tryparse(Int, x)), lines))
-            result = tryparse.(Int, lines)
-        elseif isnothing(findfirst(x -> isnothing(tryparse(Float64, x)), lines))
-            result = tryparse.(Float64, lines)
-        else
-            result = lines
-        end
-        if length(result) == 0
-            return nothing
-        elseif length(result) == 1
-            return first(result)
-        else
-            return result
-        end
-    end
-    nothing
-end
+ds9get(args...; kwds...) = ds9get(default_apt(), args...; kwds...)
+ds9get(apt::AccessPoint, args...; kwds...) = ds9get(apt, XPA.join_arguments(args); kwds...)
 
-@inline get(args...; kw...) = get(_apt(), args...; kw...)
-
-# Yields result as a vector of numerical values extracted from the binary
-# contents of the reply.
-@inline get(ap::AccessPoint, ::Type{Vector{T}}, args...; kw...) where {T} =
-    XPA.get(Vector{T}, ap, join_arguments(args); kw...)
+# Yields result as a vector of numerical values extracted from the binary contents of the
+# reply.
+ds9get(apt::AccessPoint, ::Type{Vector{T}}, args...; kwds...) where {T} =
+    XPA.get(Vector{T}, apt, join_arguments(args); kwds...)
 
 # Idem with given number of elements.
-@inline get(ap::AccessPoint, ::Type{Vector{T}}, dim::Integer, args...; kw...) where {T} =
-    XPA.get(Vector{T}, (dim,), ap, join_arguments(args); kw...)
+ds9get(apt::AccessPoint, ::Type{Vector{T}}, dim::Integer, args...; kwds...) where {T} =
+    XPA.get(Vector{T}, (dim,), apt, join_arguments(args); kwds...)
 
-# Yields result as an array of numerical values with given dimensions
-# and extracted from the binary contents of the reply.
-@inline get(ap::AccessPoint, ::Type{Array{T,N}}, dims::NTuple{N,Integer}, args...; kw...) where {T,N} =
-    XPA.get(Array{T,N}, dims, ap, join_arguments(args); kw...)
+# Yields result as an array of numerical values with given dimensions and extracted from the
+# binary contents of the reply.
+ds9get(apt::AccessPoint, ::Type{Array{T,N}}, dims::NTuple{N,Integer}, args...; kwds...) where {T,N} =
+    XPA.get(Array{T,N}, dims, apt, join_arguments(args); kwds...)
 
-# Idem but Array number of dimensions not specified.
-@inline get(ap::AccessPoint, ::Type{Array{T}}, dims::NTuple{N,Integer}, args...; kw...) where {T,N} =
-    get(ap, Array{T,N}, dims, args...; kw...)
+# Idem but number of dimensions not specified.
+ds9get(apt::AccessPoint, ::Type{Array{T}}, dims::NTuple{N,Integer}, args...; kwds...) where {T,N} =
+    ds9get(apt, Array{T,N}, dims, args...; kwds...)
 
 # Yields result as a single string.
-@inline get(ap::AccessPoint, ::Type{String}, args...; kw...) =
-    XPA.get(String, ap, join_arguments(args); kw...)
+ds9get(apt::AccessPoint, ::Type{String}, args...; kwds...) =
+    XPA.get(String, apt, join_arguments(args); kwds...)
 
-# Yields result as a vector of strings split out of the textual contents of the
-# reply.
-@inline get(ap::AccessPoint, ::Type{Vector{String}}, args...; delim = isspace, keepempty::Bool=false, kw...) =
-    split(chomp(get(ap, String, args...; kw...)), delim; keepempty=keepempty)
+# Yields result as a vector of strings split out of the textual contents of the reply.
+ds9get(apt::AccessPoint, ::Type{Vector{String}}, args...; delim = isspace, keepempty::Bool=false, kwds...) =
+    split(chomp(ds9get(apt, String, args...; kwds...)), delim; keepempty=keepempty)
 
-# Yields result as a tuple of strings split out of the textual contents of the
-# reply.
-@inline get(ap::AccessPoint, ::Type{TupleOf{String}}, args...; kw...) =
-    Tuple(get(ap, Vector{String}, args...; kw...))
+# Yields result as a tuple of strings split out of the textual contents of the reply.
+ds9get(apt::AccessPoint, ::Type{TupleOf{String}}, args...; kwds...) =
+    Tuple(ds9get(apt, Vector{String}, args...; kwds...))
 
-# Yields result as a numerical value parsed from the textual contents of the
-# reply.
-@inline get(ap::AccessPoint, ::Type{T}, args...; kw...) where {T<:Real} =
-    _parse(T, get(ap, String, args...; kw...))::T
+# Yields result as a numerical value parsed from the textual contents of the reply.
+ds9get(apt::AccessPoint, ::Type{T}, args...; kwds...) where {T<:Real} =
+    _parse(T, ds9get(apt, String, args...; kwds...))::T
 
 # Yields result as a tuple of numerical values parsed from the textual contents
 # of the reply.
-@inline get(ap::AccessPoint, ::Type{TupleOf{T}}, args...; kw...) where {T<:Real} =
-    _parse(TupleOf{T}, get(ap, String, args...; kw...))
+ds9get(apt::AccessPoint, ::Type{TupleOf{T}}, args...; kwds...) where {T<:Real} =
+    _parse(TupleOf{T}, ds9get(apt, String, args...; kwds...))
 
 _parse(::Type{T}, str::AbstractString) where {T<:Real} = parse(T, str)
 
@@ -325,114 +306,144 @@ function _parse(::Type{TupleOf{T}}, str::AbstractString) where {T<:Real}
 end
 
 """
-    get([accesspoint,] Array; endian=:native, kw...)
+    ds9get([accesspoint,] Array; endian=:native, kwds...)
 
 Returns the contents of current SAOImage/DS9 frame as an array.
 
-The keyword `endian` can be used to specify the byte order of the received
-values (see [`SAOImageDS9.byte_order`](@ref)). This method retunrs`nothing` if
-the frame is empty.
+The keyword `endian` can be used to specify the byte order of the received values (see
+[`SAOImageDS9.byte_order`](@ref)). This method returns `nothing` if the frame is empty.
+
 """
-function get(ap::AccessPoint, ::Type{Array}; endian::Union{Symbol,AbstractString}=:native, kw...)
-    T = bitpix_to_type(get(ap, Int, "fits bitpix"; kw...))
+function ds9get(apt::AccessPoint, ::Type{Array};
+                endian::Union{Symbol,AbstractString}=:native, kwds...)
+    T = bitpix_to_type(ds9get(apt, Int, "fits bitpix"; kwds...))
     if T === Nothing
         return nothing
     end
-    dims = get(ap, TupleOf{Int}, "fits size"; kw...)
-    return get(ap, Array{T}, dims, "array", byte_order(endian); kw...)
+    dims = ds9get(apt, TupleOf{Int}, "fits size"; kwds...)
+    return ds9get(apt, Array{T}, dims, "array", byte_order(endian); kwds...)
 end
 
 """
-    get([accesspoint,] VersionNumber)
+    ds9get([accesspoint,] VersionNumber)
 
 Retrieve the version of the SAOImage/DS9 program.
 """
-function get(ap::AccessPoint, ::Type{VersionNumber}; kw...)
-    str = get(ap, String, "version"; kw...)
+function ds9get(apt::AccessPoint, ::Type{VersionNumber}; kwds...)
+    str = ds9get(apt, String, "version"; kwds...)
     return VersionNumber(split(str, " ")[end])
 end
 
 """
-    set([accesspoint,] args...; data=nothing)
+    ds9set([apt,] args...; data=nothing, throwerrors=true, quiet=false, kwds...)
 
-Send a command and/or data to the SAOImage/DS9 server.
+send a command and/or data to the SAOImage/DS9 server. Optional `apt` is to specify
+another access-point to a SAOImage/DS9 server than the default one. The sent command is
+made of `args...` converted into a single string with elements of `args...`
+separated by a single space.
 
-The command is made of arguments `args...` converted into strings and merged
-with a separating spaces. Keyword `data` can be used to specify the data to
-send. An exception is thrown in case of error.
+# Keywords
 
-See also: [`SAOImageDS9.connect`](@ref), [`SAOImageDS9.get`](@ref), and
-`XPA.set`.
+* `data` is to specify the data to send.
+
+* `throwerrors` is to specify whether to throw an exception in case of error(s) in the XPA
+  reply.
+
+* `quiet` is to specify whether to not warn about any errors in the XPA reply.
+
+* `kwds...` are other keywords for `XPA.set`.
+
+See also: [`ds9connect`](@ref) and [`ds9get`](@ref).
+
 """
-function set(ap::AccessPoint, args...; data=nothing, kw...)
-    r = XPA.set(XPA.address(ap), join_arguments(args); kw...)
-    if XPA.has_errors(r)
-        m = XPA.get_message(r)
-        msg = strip(m[10:end])
-        f = findfirst("(DS9:", msg)
-        if !isnothing(f)
-            msg = msg[begin:first(f)-1]
+ds9set(args...; kwds...) = ds9set(default_apt(), args...; kwds...)
+
+ds9set(apt::AccessPoint, args...; kwds...) = ds9set(apt, join_arguments(args); kwds...)
+
+function ds9set(apt::AccessPoint, cmd::AbstractString;
+                throwerrors::Bool=true, quiet::Bool=false, kwds...)
+    r = XPA.set(apt.address, cmd; kwds...)
+    if throwerrors || !quiet
+        if length(r) == 0
+            quiet || @warn "No replies for command `$cmd`"
+        else
+            msg1 = "" # first error message if any
+            for a in r
+                if a.has_error
+                    msg = a.message
+                    quiet || @warn msg
+                    if isempty(msg1)
+                        msg1 = msg
+                    end
+                end
+            end
+            if throwerrors && !isempty(msg1)
+                error(msg1)
+            end
         end
-        @warn "XPA $msg"
-    elseif r.replies == 0
-        @warn "No replies for command `$command`"
     end
     return nothing
 end
 
-@inline set(args...; kw...) = set(_apt(), args...; kw...)
-
 """
-    set([accesspoint,] arr; mask=false, new=false, endian=:native)
+    ds9set([apt,] arr; mask=false, new=false, endian=:native, kwds...)
 
-Set the contents of the current SAOImage/DS9 frame to be array `arr`.
+set the contents of an SAOImage/DS9 frame to be array `arr`. Optional `apt` is to specify
+another access-point to a SAOImage/DS9 server than the default one.
 
 # Keywords
-- `mask`: controls the DS9 mask parameters
-- `new`: if `true`, a new frame is created;
-- `endian`: specify the byte order of `arr` (see [`byte_order`](@ref)).
+
+* `mask` is true to control the DS9 mask parameters.
+
+* `frame` specifies a frame number or is `:new` to create a new frame.
+
+* `endian` specifies the byte order of `arr` (see [`SAOImageDS9.byte_order`](@ref)).
+
+* `kwds...` are other keywords for `ds9set`.
+
 """
-function set(ap::AccessPoint, arr::DenseArray{T,N}; endian::Symbol=:native,
-    mask::Bool=false, new::Bool=false, kw...) where {T<:PixelTypes,N}
+function ds9set(apt::AccessPoint, arr::AbstractArray{T,N};
+                endian::Symbol=:native, mask::Bool=false,
+                frame=nothing, kwds...) where {T<:PixelTypes,N}
     args = String["array"]
-    new && push!(args, "new")
-    mask && push!(args, "mask")
-    set(ap, args..., _arraydescriptor(arr; endian=endian); data=arr, kw...)
+    if frame !== nothing
+        mask && throw(ArgumentError("keyword `mask` must be false if keyword `frame` is something"))
+        if frame === :new
+            push!(args, "new")
+        elseif frame isa Integer
+            ds9set(apt, "frame", frame)
+        else
+            throw(ArgumentError("keyword `frame` must be `nothing`, `:new`, or an integer"))
+        end
+    end
+    data = to_pixels(arr)
+    push!(args, array_descriptor(data; endian=endian));
+    return ds9set(apt, join(args, ' '); data=data, kwds...)
 end
+
+to_pixels(A::DenseArray{T}) where {T<:PixelTypes} = A
+to_pixels(A::AbstractArray{T}) where {T<:PixelTypes} = convert(Array{T}, A)::DenseArray{T}
 
 # Convert other pixel types.
-for (T, S) in ((Int8,   Int16),
-               (UInt16, Float32),
-               (UInt32, Float32),
-               (UInt64, Float32))
-    @eval set(ap::AccessPoint, arr::AbstractArray{$T,N}; kw...) where {N} =
-        set(ap, convert(Array{$S,N}, arr); kw...)
+for (T, S) in (Int8   => Int16,
+               UInt16 => Float32,
+               Real   => Float64,)
+    @eval to_pixels(A::AbstractArray{$T}) = convert(Array{$S}, A)::DenseArray{$S}
 end
 
-# Convert non-dense array types.
-for T in PIXELTYPES
-    @eval set(ap::AccessPoint, arr::AbstractArray{$T,N}; kw...) where {N} =
-        set(ap, convert(Array{$T,N}, arr); kw...)
-end
-
-function _arraydescriptor(arr::DenseArray{T,2};
-                          endian::Symbol=:native) where {T}
+function array_descriptor(arr::AbstractArray{T,N}; endian::Symbol=:native) where {T,N}
+    2 ≤ N ≤ 3  || throw(ArgumentError("only 2- or 3-dimensional arrays are supported"))
+    arr isa DenseArray || throw(ArgumentError("data must be a dense array"))
     bp = bitpix_of(T)
-    bp != 0 || error("unsupported data type")
-    return string("[xdim=",size(arr,1),",ydim=",size(arr,2),
-                  ",bitpix=",bp,",endian=",endian,"]")
+    iszero(bp) &&  throw(ArgumentError("unsupported pixel type `$T`"))
+    if N == 2
+        return string("[xdim=",size(arr,1),",ydim=",size(arr,2),
+                      ",bitpix=",bp,",endian=",endian,"]")
+    else
+        return string("[xdim=",size(arr,1),",ydim=",size(arr,2),",zdim=",size(arr,3),
+                      ",bitpix=",bp,",endian=",endian,"]")
+    end
 end
-
-function _arraydescriptor(arr::DenseArray{T,3};
-                          endian::Symbol=:native) where {T}
-    bp = bitpix_of(T)
-    bp != 0 || error("unsupported data type")
-    return string("[xdim=",size(arr,1),",ydim=",size(arr,2),",zdim=",size(arr,3),
-                  ",bitpix=",bp,",endian=",endian,"]")
-end
-
-_arraydescriptor(arr::DenseArray; kdws...) =
-    error("only 2D and 3D arrays are supported")
 
 """
     bitpix_of(x) -> bp
@@ -443,11 +454,11 @@ of 0 is returned if `x` is not of a supported type.
 See also [`SAOImageDS9.bitpix_to_type`](@ref).
 
 """
-bitpix_of(::DenseArray{T}) where {T} = bitpix_of(T)
-for T in PIXELTYPES
+bitpix_of(::AbstractArray{T}) where {T} = bitpix_of(T)
+for T in PIXEL_TYPES
     bp = (T <: Integer ? 8 : -8)*sizeof(T)
     @eval bitpix_of(::Type{$T}) = $bp
-    @eval bitpix_of(::$T) = $bp
+    @eval bitpix_of(::$T) = $bp # FIXME
 end
 bitpix_of(::Any) = 0
 
@@ -481,7 +492,7 @@ one of the strings (or the equivalent symbol): `"big"` for most significant byte
 `"little"` for least significant byte first or `"native"` to yield the byte order of the
 machine.
 
-See also [`SAOImageDS9.get`](@ref), [`SAOImageDS9.set`](@ref).
+See also [`ds9get`](@ref), [`ds9set`](@ref).
 
 """
 function byte_order(endian::Symbol)
@@ -507,86 +518,86 @@ byte_order(endian::AbstractString) = byte_order(Symbol(endian))
 # DRAWING
 
 """
-    draw([accesspoint,] args...; kwds...)
+    ds9draw([apt,] args...; kwds...)
 
 Draws something in SAOImage/DS9 application.
 
 The specific operation depends on the type of the arguments.
-"""
-draw(args...; kwds...) = draw(_apt(), args; kwds...)
-draw(ap::AccessPoint, ::Tuple{}; kwds...) = nothing
-draw(ap::AccessPoint, ::T; kwds...) where {T} = error("unexpected type of argument(s): $T")
 
 """
-    draw([accesspoint,] img::AbstractMatrix; kwds...)
+ds9draw(args...; kwds...) = ds9draw(default_apt(), args...; kwds...)
+ds9draw(apt::AccessPoint, ::Tuple{}; kwds...) = nothing
+ds9draw(apt::AccessPoint, ::T; kwds...) where {T} = error("unexpected type of argument(s): $T")
+
+"""
+    ds9draw([apt,] img::AbstractMatrix; kwds...)
 
 Displays image `img` (a 2-dimensional Julia array) in SAOImage/DS9.
 
 # Keywords
-- `frame`: selects the given frame number;
+- `frame`: to select a given frame number, or `:new` to draw image in a new frame;
 - `cmap`: uses the named colormap;
 - `zoom`: fixes the zoom factor;
 - `min` & `max`: fix the scale limits.
 """
-function draw(ap::AccessPoint, img::AbstractMatrix;
-    min::Union{Real,Nothing}=nothing, max::Union{Real,Nothing}=nothing,
-    cmap=nothing, frame=nothing, zoom=nothing)
-    # FIXME: pack all commands into a single one.
-    frame === nothing || set(ap, "frame", frame)
-    zoom === nothing || set(ap, "zoom to", zoom)
-    set(ap, img)
+function ds9draw(apt::AccessPoint, img::AbstractMatrix;
+                 min::Union{Real,Nothing}=nothing,
+                 max::Union{Real,Nothing}=nothing,
+                 cmap=nothing, zoom=nothing, kwds...)
+    zoom === nothing || ds9set(apt, "zoom to", zoom)
+    ds9set(apt, img; kwds...)
     if min !== nothing || max !== nothing
-        set(ap, "scale limits", limits(img, min, max)...)
+        ds9set(apt, "scale limits", limits(img, min, max)...)
     end
-    cmap === nothing || set(ap, "cmap", cmap)
+    cmap === nothing || ds9set(apt, "cmap", cmap)
     return nothing
 end
 
-# For multiple points/circles/... we just send a command to SAOImage/DS9 for
-# each item to draw.  Packing multiple commands (separated by semi-columns)
-# does not really speed-up things and is more complicated because there is a
-# limit to the total length of an XPA command (given by XPA.SZ_LINE I guess).
+# For multiple points/circles/... we just send a command to SAOImage/DS9 for each item to
+# draw. Packing multiple commands (separated by semi-columns) does not really speed-up
+# things and is more complicated because there is a limit to the total length of an XPA
+# command (given by `XPA.SZ_LINE`).
 
 """
-    draw([accesspoint,] pnt; kwds...)
+    ds9draw([accesspoint,] pnt; kwds...)
 
 Draw `pnt` as point(s) in SAOImage/DS9.
 
 `pnt` can be a `Point`, an array, or a tuple of `Point`'s.
 """
-draw(ap::AccessPoint, A::Point; kwds...) = _draw(ap, _region(Val(:point), kwds), A)
-function draw(ap::AccessPoint, A::Union{Tuple{Vararg{Point}},
-    AbstractArray{<:Point}}; kw...)
-    cmd = _region(Val(:point), kw)
+ds9draw(apt::AccessPoint, A::Point; kwds...) = _draw(apt, _region(Val(:point), kwds), A)
+function ds9draw(apt::AccessPoint, A::Union{Tuple{Vararg{Point}},
+    AbstractArray{<:Point}}; kwds...)
+    cmd = _region(Val(:point), kwds)
     for i in eachindex(A)
-        _draw(ap::AccessPoint, cmd, A[i])
+        _draw(apt::AccessPoint, cmd, A[i])
     end
 end
 
 """
-    draw([accesspoint,] box; kwds...)
+    ds9draw([accesspoint,] box; kwds...)
 
 Draws `box` as rectangle(s) in SAOImage/DS9.
 
 box` can be a `BoundingBox`, an array, or a tuple of `BoundingBox`'es.
 """
-draw(ap::AccessPoint, A::BoundingBox; kwds...) = _draw(ap, _region(Val(:polygon), kwds), A)
-function draw(ap::AccessPoint, A::Union{Tuple{Vararg{BoundingBox}},
-    AbstractArray{<:BoundingBox}}; kw...)
-    cmd = _region(Val(:polygon), kw)
+ds9draw(apt::AccessPoint, A::BoundingBox; kwds...) = _draw(apt, _region(Val(:polygon), kwds), A)
+function ds9draw(apt::AccessPoint, A::Union{Tuple{Vararg{BoundingBox}},
+    AbstractArray{<:BoundingBox}}; kwds...)
+    cmd = _region(Val(:polygon), kwds)
     for i in eachindex(A)
-        _draw(ap, cmd, A[i])
+        _draw(apt, cmd, A[i])
     end
 end
 
-function _draw(ap::AccessPoint, cmd::NTuple{2,AbstractString}, A::Point)
-    set(ap, cmd[1], A.x, A.y, cmd[2])
+function _draw(apt::AccessPoint, cmd::NTuple{2,AbstractString}, A::Point)
+    ds9set(apt, cmd[1], A.x, A.y, cmd[2])
     nothing
 end
 
-function _draw(ap::AccessPoint, cmd::NTuple{2,AbstractString}, A::BoundingBox)
+function _draw(apt::AccessPoint, cmd::NTuple{2,AbstractString}, A::BoundingBox)
     x0, x1, y0, y1 = Tuple(A)
-    set(ap, cmd[1], x0, y0, x1, y0, x1, y1, x0, y1, x0, y0, cmd[2])
+    ds9set(apt, cmd[1], x0, y0, x1, y0, x1, y1, x0, y1, x0, y0, cmd[2])
     nothing
 end
 
@@ -619,72 +630,74 @@ for id in (:circle, :ellipse, :box, :polygon, :point, :line,
 end
 
 """
-    message([accesspoint,] message; cancel=false)
+    ds9message([apt,] text; cancel=false)
 
-Display a dialog with the given `message`.
+Display a dialog with a message given by `text`.
 
-If `cancel=true`, a *Cancel* button is added to the dialog: in that case, the
-return value is `true` or `false` depending on the button pressed by the user.
+If `cancel=true`, a *Cancel* button is added to the dialog: in that case, the return value
+is `true` or `false` depending on the button pressed by the user.
+
 """
-message(msg::AbstractString; kwds...) =  message(_apt(), msg; kwds...)
-function message(ap::AccessPoint, msg::AbstractString; cancel::Bool = false)
+ds9message(msg::AbstractString; kwds...) =  ds9message(default_apt(), msg; kwds...)
+function ds9message(apt::AccessPoint, msg::AbstractString; cancel::Bool = false)
     btn = (cancel ? "okcancel" : "ok")
     cmd = "analysis message $btn {$msg}"
-    tryparse(Int, XPA.get(String, ap, cmd)) == 1
+    return tryparse(Int, XPA.get(String, apt, cmd)) == 1
 end
 
 """
-    select([accesspoint]; text="", cancel=false, coords=:image, event=:button)
+    ds9cursor([apt]; text="", cancel=false, coords=:image, event=:button)
 
-Returns the position selected by the user.
+Returns the position of the mouse cursor in SAOImage/DS9 interactively chosen by the user.
+The function returns the tuple `(key, x, y)` or `(key, value)`, where `key` is the key
+pressed, `(x, y)` are the coordinates of the point selected, and `value` the corresponding
+value.
 
 # Keywords
-- `text`: a dialog message with the given text is displayed first;
-- `cancel`: if `true`, the dialog message will allow the user to cancel the
-  operation (in this case this function returns `nothing`);
-- `event`: the type of event to capture the cursor position, as a symbol:
-  `:button`, `:key`, `:any`
-- `coords`: the type of coordinates to return as a symbol: `:image`,
-  `:physical`, `:fk5`, `:galactic`; if set to `:data`, this function returns
-  the value of the pixel, instead of its coordinates
 
-The function returns the tuple `(key, x, y)` or `(key, value)`, where `key` is the key
-pressed, `(x, y)` are the coordinates of the point selected, and `value` the
-corresponding value.
+- `text` specifies a message to be displayed in a dialog first.
+
+- `cancel` specifies whether the dialog message will let the user cancel the operation (in
+  this case this function returns `nothing`).
+
+- `event` is the type of event to capture the cursor position, one of `:button`, `:key`, or
+  `:any`.
+
+- `coords` is the type of coordinates to return, one of `:data`, `:image`, `:physical`,
+  `:fk5`, or `galactic` (as a string or as a symbol). If set to `data`, this function
+  returns the value of the pixel, instead of its coordinates.
+
 """
-function select(ap::AccessPoint=_apt(); text::AbstractString="", cancel::Bool=false,
-    coords=:image, event=:button)
-    if event ∉ (:button, :key, :any)
-        error("Unknown event type $event")
-    end
-    if length(text) > 0
-        message(ap, text; cancel=cancel) || return nothing
+function ds9cursor(apt::Union{AbstractString,AccessPoint}=default_apt();
+                   text::AbstractString="", cancel::Bool=false,
+                   coords::Symbol=:image, event::Symbol=:button)
+    event ∈ (:button, :key, :any) || throw(ArgumentError(
+        "unknown event type `$(repr(event))`, must be one of `:button`, `:key`, or `:any`"))
+    if isempty(text)
+        XPA.set(apt, "raise")
     else
-        XPA.set(XPA.address(ap), "raise")
+        ds9message(apt, text; cancel=cancel) || return nothing
     end
-    reply = XPA.get(ap, "imexam $event $(coords !== :data ? "coordinate " : "") $coords")
-    data = split(XPA.get_data(String, reply))
-    if event != :button
-        key = data[1]
-        p = parse.(Float64, @view data[2:end])
+    cmd = string("iexam ", event, (coords === :data ? " " : " coordinate "), coords)
+    words = split(XPA.get(String, apt, cmd))
+    if event === :button
+        return ("<1>", parse.(Float64, words)...)
     else
-        key = "<1>"
-        p = parse.(Float64, data)
+        return (words[1], parse.(Float64, @view words[2:end])...)
     end
-    (key, p...)
 end
 
-function cube(; apt=_apt())
+function cube(; apt=default_apt())
     get_and_parse(Int, apt, "cube")
 end
-function cube(z::Integer; apt=_apt())
+function cube(z::Integer; apt=default_apt())
     XPA.set(apt, "cube $z")
 end
 
-function cube_interval(; apt=_apt())
+function cube_interval(; apt=default_apt())
     get_and_parse(Float64, apt, "cube interval")
 end
-function cube_interval(dt::Real; apt=_apt())
+function cube_interval(dt::Real; apt=default_apt())
     XPA.set(apt, "cube interval $dt")
 end
 
@@ -802,28 +815,32 @@ valid_maximum(::typeof(isfinite), A::AbstractRange{<:Real}) = maximum(A)
 
 
 """
-    ds9wcs([access_point]; useheader=true)
+    ds9wcs([apt]; useheader=true)
 
-Return the WCS transformation of the current image in the DS9 window.
+Return the cards of the FITS header defining the WCS transformation of current
+SAOImage/DS9 frame.
 
-# Keyword Arguments
-- `useheader`: if `true`, the WCS is extracted from the FITS header. If
-  `false`, the WCS is extracted from the DS9 window.
+# Keywords
+
+- `useheader` specifies whether to extract WCS transformation from the FITS header.
+  Otherwise, the WCS transformation is extracted from the result of the `"wcs save"`
+  command.
+
 """
-function ds9wcs(ap::AccessPoint=_apt(); useheader=true)
-    wcskeys = r"^(WCSAXES |CRPIX.  |CRVAL.  |CTYPE.  |CDELT.  |PC._.   |CD._.   |PV._.   |RADESYS |LATPOLE |LONPOLE |CUNIT.  )="
+function ds9wcs(apt::Union{AbstractString,AccessPoint} = default_apt();
+                useheader::Bool = true)
+    wcskeys = r"^(WCSAXES|C(RPIX|RVAL|TYPE|DELT|UNIT)[1-9] |(PC|CD|PV)[1-9]_[1-9]  |RADESYS|LATPOLE|LONPOLE) ="
     if useheader
-        reply = XPA.get(ap, "fits header")
-        header = XPA.get_data(String, reply)
+        header = XPA.get(String, apt, "fits header")
     else
         path = tempname()
-        reply = XPA.set(XPA.address(ap), "wcs save $path")
+        XPA.set(apt, "wcs save", path)
         header = open(path) do f
             read(f, String)
         end
         rm(path, force=true)
     end
-    filter(startswith(wcskeys), split(header, "\n"))
+    return filter(startswith(wcskeys), split(header, "\n"))
 end
 
 """
@@ -857,7 +874,7 @@ types, depending on the keyword. The followin rules applies
 - properties consisting of multiple elements, such as `:dash`, `:line`, and
   `:point` are returned as tuples.
 """
-function ds9getregions(ap::AccessPoint, name::String; coords=:image, selected=false)
+function ds9getregions(apt::AccessPoint, name::String; coords=:image, selected=false)
     function parsevalue(prop, value)
         r = tryparse(Int, value)
         if r === nothing
@@ -908,7 +925,7 @@ function ds9getregions(ap::AccessPoint, name::String; coords=:image, selected=fa
     # Query DS9 for the regions and the selections
     group = (name != "") ? "-group $name" : ""
     sel = selected ? "selected" : ""
-    reply = XPA.get(ap, "regions $sel -format ds9 $group -system $coords")
+    reply = XPA.get(apt, "regions $sel -format ds9 $group -system $coords")
     regs = split(XPA.get_data(String, reply), "\n")
     # Extract the global settings
     global_lines = filter(line -> startswith(line, "global "), regs)
@@ -932,7 +949,10 @@ function ds9getregions(ap::AccessPoint, name::String; coords=:image, selected=fa
     end
     return regions
 end
-@inline ds9getregions(name::String=""; kw...) = ds9getregions(_apt(), name; kw...)
+@inline ds9getregions(name::String=""; kwds...) = ds9getregions(default_apt(), name; kwds...)
 
+function __init__()
+    ds9disconnect()
+end
 
 end # module